Marck

+++ app/Http/Controllers/UserController.php
@@ -13,6 +13,7 @@ use App\Enums\UserRole;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Exception;
+use Illuminate\Support\Facades\DB;
 class UserController extends Controller
 {
     public function index()
@@ -45,13 +46,13 @@ class UserController extends Controller
         return back()->withErrors(['email' => 'Las credenciales no coinciden'])->withInput();
     }
 
-   protected function redirectToRoleDashboard(UserRole $role)
-{
-    $route = $role->dashboardRoute(); // Using the enum method we'll create
-    
-    return redirect()->route($route)
-        ->with('success', 'Sesión iniciada correctamente');
-}
+    protected function redirectToRoleDashboard(UserRole $role)
+    {
+        $route = $role->dashboardRoute(); // Using the enum method we'll create
+
+        return redirect()->route($route)
+            ->with('success', 'Sesión iniciada correctamente');
+    }
 
     public function logout()
     {
@@ -78,50 +79,31 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
-
         try {
-
             $messages = [
                 'name.required' => 'El nombre es obligatorio.',
                 'name.string' => 'El nombre debe ser texto.',
                 'name.max' => 'El nombre no debe exceder los 100 caracteres.',
-
                 'lastname.required' => 'El apellido es obligatorio.',
                 'lastname.string' => 'El apellido debe ser texto.',
                 'lastname.max' => 'El apellido no debe exceder los 100 caracteres.',
-
                 'dni.required' => 'El DNI es obligatorio.',
                 'dni.string' => 'El DNI debe ser texto.',
                 'dni.size' => 'El DNI debe tener exactamente 8 caracteres.',
-
                 'phone.required' => 'El teléfono es obligatorio.',
                 'phone.max' => 'El teléfono no debe exceder los 9 caracteres.',
-
                 'sexo.required' => 'El sexo es obligatorio.',
                 'sexo.in' => 'El sexo debe ser M (Masculino) o F (Femenino).',
-
                 'estado_civil.required' => 'El estado civil es obligatorio.',
                 'estado_civil.in' => 'El estado civil debe ser S (Soltero), C (Casado), D (Divorciado) o V (Viudo).',
-
                 'address.required' => 'La dirección es obligatoria.',
                 'address.string' => 'La dirección debe ser texto.',
                 'address.max' => 'La dirección no debe exceder los 200 caracteres.',
-
                 'fecha_nacimiento.required' => 'La fecha de nacimiento es obligatoria.',
                 'fecha_nacimiento.date' => 'Debe ingresar una fecha válida.',
-
-                'email.required' => 'El correo electrónico es obligatorio.',
-                'email.email' => 'Debe ingresar un correo electrónico válido.',
-                'email.unique' => 'Este correo electrónico ya está registrado.',
-
-                'password.required' => 'La contraseña es obligatoria.',
-                'password.min' => 'La contraseña debe tener al menos 6 caracteres.',
-                'password.regex' => 'La contraseña debe contener al menos una mayúscula y una minúscula.',
-
                 'photo.image' => 'El archivo debe ser una imagen.',
                 'photo.max' => 'La imagen no debe exceder los 4MB.',
                 'photo.mimes' => 'La imagen debe ser JPG, PNG o JPEG.',
-
             ];
 
             $request->validate([
@@ -133,12 +115,14 @@ class UserController extends Controller
                 'estado_civil' => 'required|in:S,C,D,V',
                 'address' => 'required|string|max:200',
                 'fecha_nacimiento' => 'required|date',
-                'email' => 'required|email|unique:users',
-                'password' => 'required|min:6|regex:/[A-Z]/|regex:/[a-z]/',
                 'photo' => 'nullable|image|max:4096|mimes:jpg,png,jpeg',
-
             ], $messages);
 
+            // Generar email automático
+            $nombre = strtolower($request->name);
+            $apellido = strtolower($request->lastname);
+            $randomNumbers = rand(1000, 9999);
+            $email = $nombre . substr($apellido, 0, 1) . '_' . $randomNumbers . '@bruning.com';
 
             // Crear Persona primero
             $persona = Persona::create([
@@ -155,18 +139,16 @@ class UserController extends Controller
                     : null
             ]);
 
-            Log::info("info ", ['persona' => $persona]);
-            // Crear User asociado
+            // Crear User asociado con valores fijos
             $user = User::create([
                 'persona_id' => $persona->persona_id,
-                'email' => $request->email,
-                'password' => Hash::make($request->password),
-                'rol' => 'admin',
+                'email' => $email,
+                'password' => Hash::make('password'), // Contraseña fija
+                'rol' => 'admin', // Rol fijo como admin
                 'estado' => true
             ]);
 
-            Log::info("info ", ['persona' => $user]);
-            return redirect()->route('users.buscar')->with('success', 'Usuario creado exitosamente.');
+            return redirect()->route('users.buscar')->with('success', 'Usuario creado exitosamente. Email generado: ' . $email);
         } catch (ValidationException $e) {
             $errorMessage = implode('<br>', $e->validator->errors()->all());
             return redirect()->back()
@@ -174,11 +156,9 @@ class UserController extends Controller
                 ->with('error', $errorMessage);
         } catch (Exception $e) {
             Log::error("Error storing: " . $e->getMessage());
-
             return redirect()->back()->with('error', 'Error: ' . 'Hubo un error. Porfavor, pruebe denuevo');
         }
     }
-
     public function edit($id)
     {
         $user = User::with('persona')->findOrFail($id);
@@ -247,76 +227,102 @@ class UserController extends Controller
         }
             */
 
+
     public function update(Request $request, $user_id)
     {
-        $user = User::with('persona')->findOrFail($user_id);
+        try {
+            $user = User::with('persona')->findOrFail($user_id);
 
-        $request->validate([
-            'name' => 'required|string|max:100',
-            'lastname' => 'required|string|max:100',
-            'dni' => 'required|string|size:8',
-            'phone' => 'required|string|max:15',
-            'sexo' => 'required|in:M,F',
-            'estado_civil' => 'required|in:S,C,D,V',
-            'address' => 'required|string|max:200',
-            'fecha_nacimiento' => 'required|date',
-            'email' => 'required|email',
-            'rol' => 'required|in:' . implode(',', UserRole::values()),
-            'photo' => 'nullable|image|max:2048|mimes:jpg,png,jpeg',
-        ]);
+            $request->validate([
+                'edit_name' => 'required|string|max:100',
+                'edit_lastname' => 'required|string|max:100',
+                'edit_dni' => 'required|string|size:8',
+                'edit_phone' => 'required|string|max:15',
+                'edit_sexo' => 'required|in:M,F',
+                'edit_estado_civil' => 'required|in:S,C,D,V',
+                'edit_address' => 'required|string|max:200',
+                'edit_fecha_nacimiento' => 'required|date',
+                'edit_email' => 'required|email|unique:users,email,' . $user->user_id . ',user_id',
+                'edit_photo' => 'nullable|image|max:2048|mimes:jpg,png,jpeg',
+            ]);
 
-        // Actualizar Persona
-        $personaData = [
-            'name' => $request->name,
-            'lastname' => $request->lastname,
-            'dni' => $request->dni,
-            'phone' => $request->phone,
-            'sexo' => $request->sexo,
-            'estado_civil' => $request->estado_civil,
-            'address' => $request->address,
-            'fecha_nacimiento' => $request->fecha_nacimiento
-        ];
+            // Actualizar Persona
+            $personaData = [
+                'name' => $request->edit_name,
+                'lastname' => $request->edit_lastname,
+                'dni' => $request->edit_dni,
+                'phone' => $request->edit_phone,
+                'sexo' => $request->edit_sexo,
+                'estado_civil' => $request->edit_estado_civil,
+                'address' => $request->edit_address,
+                'fecha_nacimiento' => $request->edit_fecha_nacimiento
+            ];
 
-        if ($request->hasFile('photo')) {
-            // Eliminar foto anterior si existe
-            if ($user->persona->photo) {
-                Storage::disk('public')->delete($user->persona->photo);
+            if ($request->hasFile('edit_photo')) {
+                if ($user->persona->photo) {
+                    Storage::disk('public')->delete($user->persona->photo);
+                }
+                $personaData['photo'] = $request->file('edit_photo')->store('profile_photos', 'public');
             }
-            $personaData['photo'] = $request->file('photo')->store('profile_photos', 'public');
+
+            $user->persona->update($personaData);
+
+            // Actualizar User (email)
+            $user->update([
+                'email' => $request->edit_email
+            ]);
+
+            return redirect()->route('users.buscar')->with('success-update', 'Usuario actualizado con éxito');
+        } catch (Exception $e) {
+            Log::error("Error updating: " . $e->getMessage());
+            return redirect()->back()->with('error', 'Error: ' . 'Hubo un error. Porfavor, pruebe denuevo');
         }
-
-        $user->persona->update($personaData);
-
-        // Actualizar User
-        $user->update([
-            'email' => $request->email,
-            'rol' => $request->rol
-        ]);
-
-        return redirect()->route('users.buscar')->with('success-update', 'Usuario actualizado con éxito');
     }
 
 
     public function destroy($user_id)
     {
 
-        $user = User::with('persona')->findOrFail($user_id);
+       DB::beginTransaction();
 
-        // Eliminar foto si existe
-        if ($user->persona->photo) {
+    try {
+        $user = User::with(['persona', 'docente', 'secretaria', 'tutor'])->findOrFail($user_id);
+
+        // 1. Eliminar primero los modelos relacionados más específicos
+        if ($user->docente) {
+            $user->docente->delete();
+        }
+
+        if ($user->secretaria) {
+            $user->secretaria->delete();
+        }
+
+        if ($user->tutor) {
+            $user->tutor->delete();
+        }
+
+        // 2. Eliminar archivos asociados
+        if ($user->persona && $user->persona->photo) {
             Storage::disk('public')->delete($user->persona->photo);
         }
 
-        // Eliminar primero los registros relacionados (si existen)
-        optional($user->docente)->delete();
-        optional($user->secretaria)->delete();
-        optional($user->tutor)->delete();
-
-        // Eliminar persona y usuario
-        $user->persona->delete();
+        // 3. Eliminar el usuario (que tiene la FK a persona)
         $user->delete();
 
-        return redirect()->route('users')->with('success-destroy', 'Usuario eliminado exitosamente.');
+        // 4. Finalmente eliminar la persona
+        if ($user->persona) {
+            $user->persona->delete();
+        }
+
+        DB::commit();
+
+        return redirect()->route('users.buscar')->with('success-destroy', 'Usuario eliminado exitosamente.');
+    } catch (\Exception $e) {
+        DB::rollBack();
+        Log::error('Error al eliminar usuario: ' . $e->getMessage());
+        
+        return redirect()->back()->with('error', 'No se pudo eliminar el usuario debido a relaciones existentes. Error: ' . $e->getMessage());
+    }
     }
 
     public function edit_user()
@@ -350,7 +356,7 @@ class UserController extends Controller
             'address',
         ]);
 
-        if ($request->hasFile('photo')) {
+        if ($request->hasFile('photo') && Storage::disk('public')->exists($user->persona->photo)) {
             if ($user->persona->photo) {
                 Storage::disk('public')->delete($user->persona->photo);
             }
